requirements.txt
requests
beautifulsoup4==4.13.4
blinker==1.9.0
boto3==1.38.34
botocore==1.38.34
passlib==1.7.4
psycopg2-binary==2.9.10
flask-migrate==4.0.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
Flask==3.1.1
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
greenlet==3.2.3
gunicorn==23.0.0
idna==3.10
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
jmespath==1.0.1
MarkupSafe==3.0.2
packaging==25.0
pluggy==1.6.0
Pygments==2.19.1
pytest==8.4.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
requests==2.32.4
s3transfer==0.13.0
six==1.17.0
soupsieve==2.7
SQLAlchemy==2.0.41
typing_extensions==4.14.0
urllib3==2.4.0
Werkzeug==3.1.3
WTForms==3.2.1
flask_cors==4.0.0
numpy
openai
confluent-kafka
selenium
deep-translator
cachetools
